Java Doc added everywhere

+++ src/main/java/group5/hotelms/model/Hotel.java
@@ -18,34 +18,65 @@ public class Hotel implements Serializable {
         this.rooms = rooms;
     }
 
-    public long getId() {
+    /**
+     *
+     * @return id
+     */
+    public int getId() {
         return id;
     }
 
+    /**
+     *
+     * @return name
+     */
     public String getName() {
         return name;
     }
 
+    /**
+     *
+     * @return city
+     */
     public City getCity() {
         return city;
     }
 
+    /**
+     *
+     * @return rooms
+     */
     public Set<Room> getRooms() {
         return rooms;
     }
 
+    /**
+     * Setter for name
+     * @param name
+     */
     public void setName(String name) {
         this.name = name;
     }
 
+    /**
+     * Setter for city
+     * @param city
+     */
     public void setCity(City city) {
         this.city = city;
     }
 
+    /**
+     * Setter for rooms
+     * @param rooms
+     */
     public void setRooms(Set<Room> rooms) {
         this.rooms = rooms;
     }
 
+    /**
+     * Overrides equals
+     */
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
@@ -56,11 +87,17 @@ public class Hotel implements Serializable {
         return id == hotel.id;
     }
 
+    /**
+     * Overrides hashcode
+     */
     @Override
     public int hashCode() {
         return (int)id;
     }
 
+    /**
+     * Overrides toString
+     */
     @Override
     public String toString() {
         return "Hotel{" +
